build: use custom feeds
Rather than relying on jekyll-feed, use a custom feed. This allows us to
generate feeds based on categories, which is a crucial feature to us.
We want filters for remote "planet"-style aggregators.

Upstream jekyll-feed did not implement this even after 2 years of time
after the first user report. Official statement is along the lines of
"not sure people need this".

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>